restructure map components add missing interactions for center map
