+++ app/src/components/map/olMapHelpers.js
@@ -4,7 +4,6 @@ import Cluster from 'ol/source/Cluster';
 import VectorSource from 'ol/source/Vector';
 import { LineString, Point, Polygon } from 'ol/geom';
 import { Vector as VectorLayer } from 'ol/layer';
-import { createEmpty, extend, getWidth } from 'ol/extent';
 import monotoneChainConvexHull from 'monotone-chain-convex-hull';
 import store from '@/store';
 
@@ -21,6 +20,7 @@ import {
 } from 'ol/style';
 import { indicatorClassesIcons } from '../../config/trilateral';
 import { getColor } from './olMapColors';
+import { getHoverFeature } from './centerMapInteractions';
 
 const onStylesLoaded = [];
 
@@ -57,6 +57,7 @@ const indicatorClassesStyles = Object.keys(indicatorClassesIcons).reduce((acc, k
 // eslint-disable-next-line import/prefer-default-export
 export function createLayerFromConfig(config) {
   let source;
+  console.log(config);
   if (config.protocol === 'xyz') {
     source = new XYZSource({
       attributions: config.attribution,
@@ -214,17 +215,13 @@ function clusterCircleStyle(cluster, resolution) {
   }, []);
 }
 
-let hoverFeature;
-let clickInteraction;
-let pointermoveInteraction;
-
 /**
  * Style for convex hulls of clusters, activated on hover.
  * @param {Feature} cluster The cluster feature.
  * @return {Style} Polygon style for the convex hull of the cluster.
  */
-function clusterHullStyle(cluster) {
-  if (cluster !== hoverFeature) {
+export function clusterHullStyle(cluster) {
+  if (cluster !== getHoverFeature()) {
     return;
   }
   const originalFeatures = cluster.get('features');
@@ -317,63 +314,6 @@ export function createIndicatorFeatureLayers(indicators, vm) {
   return [clusterHulls, clusters, clusterCircles];
 }
 
-/**
- *
- * @param {Object} map ol map
- * @param {Object} vm vue instance
- */
-// eslint-disable-next-line import/prefer-default-export
-export function initInteractions(map, vm) {
-  const clusters = map.getLayers().getArray().find((l) => l.get('name') === 'clusters');
-  const clusterHulls = map.getLayers().getArray().find((l) => l.get('name') === 'clusterHulls');
-  const clusterCircles = map.getLayers().getArray().find((l) => l.get('name') === 'clusterCircles');
-  pointermoveInteraction = (event) => {
-    clusters.getFeatures(event.pixel).then((features) => {
-      if (features[0] !== hoverFeature) {
-        // Display the convex hull on hover.
-        // eslint-disable-next-line prefer-destructuring
-        hoverFeature = features[0];
-        clusterHulls.setStyle(clusterHullStyle);
-        // Change the cursor style to indicate that the cluster is clickable.
-        // eslint-disable-next-line no-param-reassign
-        map.getTargetElement().style.cursor = hoverFeature && hoverFeature.get('features').length > 1
-          ? 'pointer'
-          : '';
-      }
-    });
-  };
-  map.on('pointermove', pointermoveInteraction);
-
-  clickInteraction = (event) => {
-    clusters.getFeatures(event.pixel).then((features) => {
-      if (features.length > 0) {
-        const clusterMembers = features[0].get('features');
-        if (clusterMembers.length > 1) {
-          // Calculate the extent of the cluster members.
-          const extent = createEmpty();
-          clusterMembers.forEach((feature) => extend(extent, feature.getGeometry().getExtent()));
-          const view = map.getView();
-          const resolution = map.getView().getResolution();
-          if (
-            view.getZoom() === view.getMaxZoom()
-            || (getWidth(extent) < resolution && getWidth(extent) < resolution)
-          ) {
-            // Show an expanded view of the cluster members.
-            // eslint-disable-next-line prefer-destructuring
-            clickFeature = features[0];
-            clickResolution = resolution;
-            clusterCircles.setStyle(clusterCircleStyle.bind(vm));
-          } else {
-            // Zoom to the extent of the cluster members.
-            view.fit(extent, { duration: 500, padding: [50, 50, 50, 50] });
-          }
-        }
-      }
-    });
-  };
-
-  map.on('click', clickInteraction);
-}
 
 /**
  * cleanup of cluster interactions (hover + click)
